requirements.txt
ultralytics==8.2.32
pillow==10.3.0
opencv-python==4.10.0.82
numpy==1.26.4
comet_ml==3.43.1
python-dotenv
google-cloud-speech==2.26.0
pyaudio==0.2.14
SpeechRecognition==3.10.4
google-cloud-texttospeech==2.16.3
pygame==2.5.2
openai==1.34.0
pyyaml==6.0.1